Added the coconut rice recipe and its picture.
Clears at the title. I was thinking of how could I handle the introduction
of new recipes without doing it through html or the frontend as I'm doing right now
and having a single page for each recipe without creating them manually and I was thinking
that it coul be done through an SQL Database but I guess I'll learn that when I get
into the backend learning path.